[hotfix][table] Extract computeCost in FlinkLogicalJoin to base class
This commit can be squashed with a following commit after code review
